PixelSpace/requirements.txt
absl-py==2.3.1
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
diffusers==0.34.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
grpcio==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lmdb==1.6.2
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pillow==11.3.0
pip==25.1
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1-2209072238
threadpoolctl==3.6.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing-extensions==4.14.0
urllib3==2.5.0
werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
